### 2017-03-04 #194 @whedon
[REVIEW]: lawn: An R client for the Turf Javascript Library for Geospatial Analysis
**Submitting author:** @jhollist (<a href="http://orcid.org/0000-0002-9254-9740">Jeffrey W Hollister</a>)
**Repository:** <a href="https://github.com/ropensci/lawn" target ="_blank">https://github.com/ropensci/lawn</a>
**Version:** v0.3.1
