### 2017-03-06 ropensci/lawn#61 @jhollist
Zenodo DOI
@sckott Heard back from @arfon on https://github.com/openjournals/joss-reviews/issues/194.  We need to link to Zenodo and push a release to get our DOI.  Since I don't have admin rights, can you take care of that?  Once the DOI is set I'll get everything else finalized.
