### 2017-03-06 ropensci/lawn#61 @sckott
sure, should the release be tied to a cran release or just tag it wherever it's at right now?
