### 2017-03-06 ropensci/lawn#61 @jhollist
That is a good question.  @arfon, any insights?  I'm thinking we can do a release for JOSS at the current version?
