### 2017-03-06 ropensci/lawn#61 @arfon
> That is a good question. @arfon, any insights? I'm thinking we can do a release for JOSS at the current version?

Ideally it should be the same version as was submitted to JOSS (Version: `v0.3.1`).
