### 2017-03-06 ropensci/lawn#61 @jhollist
OK.

So we had another commit that bumped the version since I submitted the repository to JOSS.  Can we edit the submission to match the current version 0.3.2 (minus the dev version) and then release that?
