### 2017-03-06 ropensci/lawn#61 @jhollist
I'm out of the office this afternoon but will finalize all of this tomorrow.  Thanks @sckott and @arfon !
